pyproject.toml
[tool.poetry]
name = "cvpilot"
version = "0.1.0"
description = "AI Co-pilot for CV"
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.2"
pdfplumber = "^0.11.7"
langchain = "^0.3.27"
openai = "^1.97.1"
langchain-community = "^0.3.27"
python-docx = "^1.2.0"
streamlit = {version = "1.47.0", python = "^3.11"}
google-genai = "^1.28.0"
pypdf = "^5.9.0"
tiktoken = "^0.10.0"
langchain-anthropic = "^0.3.18"
sentence-transformers = "2.7.0"
numpy = "1.26.4" 
scipy = "1.13.1" 
scikit-learn = "1.5.1"
google-generativeai = "^0.8.5"
pkginfo = "^1.10.0"
chromadb = "^1.0.16"
torch = "2.2.2"
transformers = "4.39.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
